refactor(): 把ef的仓储注册放到ef类库
