+++ src/DNIC.Erechtheion.Core/DNIC.Erechtheion.Core.csproj
@@ -7,4 +7,10 @@
 		<PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.1" />
 		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.0.1" />
 	</ItemGroup>
+
+	<ItemGroup>
+	  <Reference Include="Microsoft.Extensions.DependencyInjection.Abstractions">
+	    <HintPath>C:\Program Files\dotnet\sdk\NuGetFallbackFolder\microsoft.extensions.dependencyinjection.abstractions\2.0.0\lib\netstandard2.0\Microsoft.Extensions.DependencyInjection.Abstractions.dll</HintPath>
+	  </Reference>
+	</ItemGroup>
 </Project>
