+++ src/DNIC.Erechtheion/DependencyInjectionConfig.cs
@@ -1,12 +1,7 @@
 ﻿using DNIC.Erechtheion.Application;
 using DNIC.Erechtheion.Application.Service;
 using DNIC.Erechtheion.Core;
-using DNIC.Erechtheion.Core.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using DNIC.Erechtheion.EntityFrameworkCore;
-using DNIC.Erechtheion.Domain;
-using DNIC.Erechtheion.Domain.Repositories;
-using DNIC.Erechtheion.EntityFrameworkCore.Repositories;
 using DNIC.Erechtheion.Identity.EntityFrameworkCore;
 
 namespace DNIC.Erechtheion
@@ -15,10 +10,9 @@ namespace DNIC.Erechtheion
 	{
 		public static void AddErechtheion(this IServiceCollection services)
 		{
-			services.AddSingleton<IErechtheionConfiguration, ErechtheionConfiguration>();
-			services.AddSingleton<ISeedDataInitiator, SeedDataInitiator>();
 			services.AddSingleton<ISeedDataInitiator, IdentitySeedDataInitiator>();
-			services.AddScoped<ITopicRepository, TopicRepository>();
+
+			//注册应用服务，如果分离了就注册客户端
 			services.AddScoped<ITopicApplicationService, TopicApplicationService>();
 		}
 	}
